requirements.txt
flask>=2.3.3
slack-bolt>=1.18.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
